7/26_1

+++ src/main/java/kr/ac/dankook/ace/healthy_meal_backend/entity/DietCriterion.java
@@ -25,7 +25,7 @@ public class DietCriterion {
     private Character gender;
 
     @Column(name = "energy_kcal")
-    private Integer energyKcal;
+    private Float energyKcal;
 
     @Column(name = "protein_g")
     private Float proteinG;
